requirements.txt
bcrypt==4.0.1
bidict==0.22.1
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.2
Flask-JWT-Extended==4.4.4
Flask-SocketIO==5.3.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pydantic==1.10.7
PyJWT==2.6.0
python-dotenv==1.0.0
python-engineio==4.4.1
python-socketio==5.8.0
requests==2.30.0
SQLAlchemy==2.0.12
typing_extensions==4.5.0
urllib3==2.0.2
Werkzeug==2.3.3
